+++ R/compile.R
@@ -51,14 +51,17 @@ compile.ConservationProblem <- function(x, compressed_formulation = NA, ...) {
     is.na(compressed_formulation) ||
           assertthat::is.flag(compressed_formulation))
   # sanity checks
-  if (inherits(x$objective, "MaximumCoverageObjective") &
+  if (inherits(x$objective, c("MaximumUtilityObjective",
+                              "MaximumCoverageObjective")) &
       !is.Waiver(x$targets))
     warning("ignoring targets since the maximum coverage objective function ",
             "doesn't use targets")
   # replace waivers with defaults
   if (is.Waiver(x$objective))
     x <- add_default_objective(x)
-  if (is.Waiver(x$targets) & !inherits(x$objective, "MaximumCoverageObjective"))
+  if (is.Waiver(x$targets) & !inherits(x$objective,
+                                       c("MaximumUtilityObjective",
+                                         "MaximumCoverageObjective")))
     x <- add_default_targets(x)
   if (is.Waiver(x$decisions))
     x <- add_default_decisions(x)