+++ src/rcpp_apply_max_cover_objective.cpp
@@ -2,10 +2,8 @@
 #include "optimization_problem.h"
 
 // [[Rcpp::export]]
-bool rcpp_apply_max_cover_objective(SEXP x,
-                                           Rcpp::NumericVector abundances,
-                                           Rcpp::NumericVector costs,
-                                           double budget) {
+bool rcpp_apply_max_cover_objective(SEXP x, Rcpp::NumericVector costs,
+                                    double budget) {
   Rcpp::XPtr<OPTIMIZATIONPROBLEM> ptr = Rcpp::as<Rcpp::XPtr<OPTIMIZATIONPROBLEM>>(x);
 
   // initialize
@@ -25,7 +23,7 @@ bool rcpp_apply_max_cover_objective(SEXP x,
   ptr->_rhs.push_back(budget);
   // model sense variables
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_sense.push_back("=");
+    ptr->_sense.push_back(">=");
   ptr->_sense.push_back("<=");
   // add in small negative number to objective for planning unit variables to
   // break ties in solution and select solution with cheapest cost
@@ -34,39 +32,41 @@ bool rcpp_apply_max_cover_objective(SEXP x,
     ptr->_obj.push_back(costs[i] * cost_scale);
   if (!ptr->_compressed_formulation)
     for (std::size_t i=0; i<A_extra_ncol; ++i)
-       ptr->_obj.push_back(0.0);
-  // add in default feature weights (features weighted equally by default)
+      ptr->_obj.push_back(0.0);
+  // add in default species weights (species treated equally)
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
     ptr->_obj.push_back(1.0);
-  // add in upper bounds representing the maximum abundances for each feature
+  // add in upper and lower bounds for the decision variables representing if
+  // each species is adequately conserved
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_ub.push_back(abundances[i]);
-  // add in lower bounds as zero
-for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
+    ptr->_ub.push_back(1.0);
+  for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
     ptr->_lb.push_back(0.0);
-  // add continuous variables representing how much of each species is
-  // conserved
+  // add in binary variable types for variables representing if each species is
+  // adequately conserved
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_vtype.push_back("C");
-  // add in model matrix values for feature abundances
+    ptr->_vtype.push_back("B");
+  // add in model matrix values for species representation amounts
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
     ptr->_A_i.push_back(A_extra_nrow + i);
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
     ptr->_A_j.push_back(ptr->_number_of_planning_units + A_extra_ncol + i);
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_A_x.push_back(-1.0);
+    ptr->_A_x.push_back(-1);
+
   // add in budget constraints
-    for (std::size_t j=0; j<(ptr->_number_of_planning_units); ++j)
-      ptr->_A_i.push_back(A_extra_nrow + ptr->_number_of_features);
+  for (std::size_t j=0; j<(ptr->_number_of_planning_units); ++j)
+    ptr->_A_i.push_back(A_extra_nrow + ptr->_number_of_features);
   for (std::size_t j=0; j<(ptr->_number_of_planning_units); ++j)
     ptr->_A_j.push_back(j);
   for (std::size_t j=0; j<(ptr->_number_of_planning_units); ++j)
     ptr->_A_x.push_back(costs[j]);
+
   // add in row and col ids
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_col_ids.push_back("amount");
+    ptr->_col_ids.push_back("spp_met");
   for (std::size_t i=0; i<(ptr->_number_of_features); ++i)
-    ptr->_row_ids.push_back("spp_amount");
+    ptr->_row_ids.push_back("spp_target");
   ptr->_row_ids.push_back("budget");
   ptr->_modelsense="max";
   return true;