+++ tests/testthat/test_add_max_features_objective.R
@@ -36,11 +36,11 @@ test_that("solution (compressed formulation)", {
   skip_on_cran()
   # create data
   budget <- 4.23
-  cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 2))
+  cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
   locked_in <- 2
   locked_out <- 1
-  features <- raster::stack(raster::raster(matrix(c(2, 1, 1, 0), ncol = 2)),
-                            raster::raster(matrix(c(10, 10, 10, 10), ncol = 2)))
+  features <- raster::stack(raster::raster(matrix(c(2, 1, 1, 0), ncol = 4)),
+                            raster::raster(matrix(c(10, 10, 10, 10), ncol = 4)))
   # create problem
   p <- problem(cost, features) %>%
         add_max_features_objective(budget = budget) %>%
@@ -106,11 +106,11 @@ test_that("solution (expanded formulation)", {
   skip_on_cran()
   # create data
   budget <- 4.23
-  cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 2))
+  cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
   locked_in <- 2
   locked_out <- 1
-  features <- raster::stack(raster::raster(matrix(c(2, 1, 1, 0), ncol = 2)),
-                            raster::raster(matrix(c(10, 10, 10, 10), ncol = 2)))
+  features <- raster::stack(raster::raster(matrix(c(2, 1, 1, 0), ncol = 4)),
+                            raster::raster(matrix(c(10, 10, 10, 10), ncol = 4)))
   # create problem
   p <- problem(cost, features) %>%
         add_max_features_objective(budget = budget) %>%